Restructure how harness runs python scripts
