+++ .gitignore
@@ -1,5 +1,4 @@
-# Build artefacts
-*/build/
+# artefacts
 io/
 datasets/
 third_party/
@@ -47,3 +46,4 @@ harness/__pycache__/
 # Virtual environments
 virtualenv
 
+/fetch_by_similarity.egg-info/
