--- run-venv
@@ -1,78 +1,0 @@
-#!/usr/bin/env bash
-# Run script for fetch-by-similarity with proper environment setup
-
-set -euo pipefail
-
-# Set OpenFHE installation path
-export CMAKE_PREFIX_PATH=~/p/openfhe-development/install
-#export LATTICA_BE_URL=http://localhost:3050
-
-# Get the directory where this script is located
-SCRIPT_DIR="$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )"
-cd "$SCRIPT_DIR"
-
-# Activate Python virtual environment
-source ./bmenv/bin/activate
-
-echo "Environment setup complete:"
-echo "  CMAKE_PREFIX_PATH=$CMAKE_PREFIX_PATH"
-echo "  Python venv: $VIRTUAL_ENV"
-echo "  Working directory: $(pwd)"
-echo ""
-
-# Function to show usage
-show_usage() {
-    echo "Usage: $0 <dataset_size> [options]"
-    echo ""
-    echo "Dataset sizes:"
-    echo "  0 - toy dataset"
-    echo "  1 - small dataset"
-    echo "  2 - medium dataset"
-    echo "  3 - large dataset"
-    echo ""
-    echo "Options:"
-    echo "  --count_only     - Run in count-only mode (no payloads)"
-    echo "  --seed <number>  - Set random seed"
-    echo "  --num_runs <n>   - Number of benchmark runs"
-    echo ""
-    echo "Examples:"
-    echo "  $0 0                           # Run toy dataset"
-    echo "  $0 0 --count_only              # Run toy dataset, count-only mode"
-    echo "  $0 1 --seed 12345 --num_runs 3 # Run small dataset with custom params"
-}
-
-# Check if any arguments provided, default to toy (0) if none
-if [ $# -eq 0 ]; then
-    echo "No arguments provided, using default: toy dataset (0)"
-    set -- 0
-fi
-
-# Check for help
-case "${1:-}" in
-    "help"|"-h"|"--help")
-        show_usage
-        exit 0
-        ;;
-esac
-
-# Determine dataset size from first argument
-SIZE=$1
-case "$SIZE" in
-    0) SIZE_NAME="toy" ;;
-    1) SIZE_NAME="small" ;;
-    2) SIZE_NAME="medium" ;;
-    3) SIZE_NAME="large" ;;
-    *) SIZE_NAME="" ;;
-esac
-
-# Clean up previous test artifacts for this specific size
-if [ -n "$SIZE_NAME" ]; then
-    echo "Cleaning up previous $SIZE_NAME dataset..."
-    rm -rf "datasets/$SIZE_NAME"
-    rm -rf "io/$SIZE_NAME"
-    echo "Cleanup complete"
-    echo ""
-fi
-
-echo "Running benchmark..."
-python harness/run_submission.py "$@"
