+++ submission/src/05_server_preprocess_dataset.py
@@ -5,8 +5,7 @@ server_preprocess_dataset.py - Upload encrypted database to server
 import sys
 import os
 
-# Add lib directory to path
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), 'lib'))
+# No path manipulation needed with proper package structure
 
 # from lattica_client import LatticaClient
